requirements.txt
asgiref==3.7.2
Django==4.2.6
python-decouple==3.8
sqlparse==0.4.4
django-ckeditor~=6.7.0